requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
scipy==1.11.4
Pillow==10.0.1
PyWavelets==1.4.1
scikit-learn==1.3.2
bitstring==4.0.2



